Added loading.tsx file for the /dashboard route to implement page-level streaming and show a fallback UI during content loading. Implemented loading skeletons to replace simple loading text, enhancing the user experience with visual feedback while the content is being fetched. Utilized Route Groups to apply loading states specifically to the /dashboard/overview route and avoid affecting other pages (/invoices, /customers). Refactored RevenueChart component to fetch its own data internally and wrapped it in Suspense with a RevenueChartSkeleton fallback for smoother loading. Moved fetchLatestInvoices() function from the main page component to the LatestInvoices component, making the component responsible for its own data fetching. Added Suspense boundaries around individual components (RevenueChart and LatestInvoices) to improve loading performance and prevent blocking the entire page. Created CardWrapper component to group the loading of multiple Card components, minimizing the popping effect and creating a more staggered loading experience. Updated the page.tsx file to use the new CardWrapper and refactored data fetching accordingly. Optimized where Suspense boundaries are placed to balance between loading speed and preventing a jarring experience due to multiple components loading asynchronously.
